pyproject.toml
[tool.poetry]
name = "taskmanager"
version = "0.1.0"
description = "SLURM task manager for molecular dynamics simulations"
authors = ["Mykyta Bobylyow <nbobylyov@gmail.com>"]
packages = [
    { include = "taskmanager", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
source = ["src/taskmanager"]
omit = ["src/taskmanager/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "pass",
    "raise ImportError"
] 